refaktoriserer brukertyper
